Allow enabling the count badge via a prop
This prop is configurable so that it can be turned on for all Refinement
components without making a custom component to override it.

The feature is turned off by default to preserve the existing behavior.

The component's state is aliased as `refinementDisplay` so that the
total can be accessed by child components. The more intuitive
`refinement` is already taken by the list item of `RefinementList`.